Split out the summary section of the shared action plan presenter and view classes.
The motivation for this is that some of places where these classes are used only
require the summary list view of the action plan. Future commits will add more
details to the action plan view class which is not required to just show the summary.
This lays the groundwork for making these future commits simpler and more efficient.
